image/requirements.txt
certbuilder~=0.14
grpcio~=1.41
Jinja2~=3.0
kopf~=1.35
kubernetes_asyncio~=18.20
kubernetes~=19.15
mypy-extensions~=0.4
protobuf~=3.19
pyyaml~=6.0
yamale~=4.0
